python-backend/requirements.txt
livekit-agents>=0.12.11
livekit-plugins-openai>=0.10.17
livekit-plugins-deepgram>=0.6.17
livekit-plugins-elevenlabs>=0.7.11
livekit-plugins-silero>=0.7.4
livekit-plugins-turn-detector
python-dotenv~=1.0
aiohttp>=3.11.11
